feat(path): add command to get path of note
The `path` command will return the path of a note with a given title if
it exists
