+++ cmd/new.go
@@ -2,9 +2,6 @@ package cmd
 
 import (
 	"fmt"
-	"io/fs"
-	"log"
-	"path/filepath"
 
 	"github.com/aadam-ali/second-brain-cli/config"
 	"github.com/spf13/cobra"
@@ -35,26 +32,6 @@ var newCmd = &cobra.Command{
 			fmt.Printf("Note already exists: %s\n", existingNoteFilepath)
 		}
 	},
-}
-
-func checkIfNoteExists(rootDir string, name string) (bool, string) {
-	pathToNote := ""
-	name = name + ".md"
-
-	err := filepath.WalkDir(rootDir, func(path string, d fs.DirEntry, err error) error {
-		if !d.IsDir() && name == d.Name() {
-			pathToNote = path
-		}
-		return nil
-	})
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	if pathToNote != "" {
-		return true, pathToNote
-	}
-	return false, ""
 }
 
 func renderStdNoteContent(title string) string {
