+++ cmd/utils.go
@@ -2,8 +2,10 @@ package cmd
 
 import (
 	"fmt"
+	"io/fs"
 	"log"
 	"os"
+	"path/filepath"
 )
 
 func constructNotePath(dir string, title string) string {
@@ -19,4 +21,24 @@ func createNote(filepath string, content string) {
 		errMsg := fmt.Sprintf("Failed to create note: %s", err)
 		log.Fatal(errMsg)
 	}
+}
+
+func checkIfNoteExists(rootDir string, name string) (bool, string) {
+	pathToNote := ""
+	name = name + ".md"
+
+	err := filepath.WalkDir(rootDir, func(path string, d fs.DirEntry, err error) error {
+		if !d.IsDir() && name == d.Name() {
+			pathToNote = path
+		}
+		return nil
+	})
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	if pathToNote != "" {
+		return true, pathToNote
+	}
+	return false, ""
 }
